story telling dialogues
I built the selection of the personas and based on that I did story telling for each persona in every stage of the game. so there is story telling after selection of the persona then after level 3 then after level 6 and at the end of the game. so generally there are 4 scenes of story telling for each persona after selection process .